requirements.txt
absl-py==0.12.0
asgiref==3.3.4
astunparse==1.6.3
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
Django==3.2.3
django-heroku==0.3.1
gast==0.3.3
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.37.1
gunicorn==20.1.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
natsort==7.1.1
networkx==2.5.1
numpy==1.20.3
oauthlib==3.1.0
opencv-python==4.5.2.52
opt-einsum==3.3.0
Pillow==8.2.0
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scipy==1.4.1
six==1.16.0
sqlparse==0.4.1
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.0
tensorflow-cpu
tensorflow-estimator==2.2.0
termcolor==1.1.0
tifffile==2021.4.8
urllib3==1.26.4
Werkzeug==2.0.1
wrapt==1.12.1
xlwt==1.3.0
